added REQ of readline and ncurses to bash added attribute of solo to bzip2 new version of libesmtp
sorcery has many bug fixes, enhanced locking, and now the ability to
simultaneously download while downloading and cast while casting in an
unlimited fashion.  If UNLIMITED is on then what cast receives on the
command line is processes in the logical order as and with as much done
in parallel as possible.  Logical sequental sequence of casting is
used when UNLIMITED is off or casting is done from a non terminal
such as by atd or crond ( at job or cronjob ).  I will probably
place a limiter on this in a few days, so that the SA can specify
how much casting they want done in parallel.