Fix broken Linux-PAM
The elegant solution of adding "64" to OPTS variable during cast
is wrong because OPTS is modified before the addition occurs. The
minimum modification is a space appended so when the spell finally
does OPTS+="64" the result is, for example, "... --libdir/usr/lib 64"
and the cast fails.
Reverted to a less elegant but explicit solution.