Bluetooth: Switch bluetooth adapters from HID to HCI mode on start
From http://times.usefulinc.com/2004/06/12-hidproxy :

One of the more annoying things about Bluetooth-enabled keyboards is
that they require software running on the host computer in order to
process the input events. This means that you cannot use them to
access BIOS screens or other parts of your computer that run before
the operating system is fully operational.

Happily, there's a really smart solution to this. Some Bluetooth
modules have a feature called "HID proxy". HID is the USB protocol
used by conventional USB mice and keyboards, and implemented in the
BIOS of all modern motherboards.

When in HID proxy mode, a Bluetooth dongle handles the input from
Bluetooth keyboards and mice itself, and turns it into HID events
to send via USB. When you plug such a dongle in, the computer
believes it's talking to a mouse and keyboard.

The dongle can then be switched into Bluetooth "HCI mode" again
when the operating system has fully booted, and the mouse and
keyboard will reattach themselves.

The Bluetooth tools for Linux, BlueZ, include such a command: hid2hci.