Fix the httpd init.d script
The httpd init.d script relied on the /lib/lsb/init-functions to
provide the status() function but it did not work. It turns out
that the lsb function expect the SERV name to match the name of
the running process which is not true for httpd: SERV is
/usr/sbin/apachectl and processes are /usr/sbin/httpd.

Due to the above issue, the httpd server did not get restarted
on update, or on PHP update. Quite annoying at the very least.

Fixed by writing a new status() function for the httpd init.d
script.